state inpagenav functionality
