+++ src/js/components/state/StatePage.jsx
@@ -3,10 +3,10 @@
  * Created by Lizzie Salita 5/2/18
  */
 
-import React from 'react';
+import React, { useState, useEffect } from 'react';
 import PropTypes from 'prop-types';
+import { find, throttle } from 'lodash';
 import { ShareIcon, FiscalYearPicker } from 'data-transparency-ui';
-
 import { statePageMetaTags } from 'helpers/metaTagHelper';
 import { currentFiscalYear, earliestFiscalYear, getFiscalYearsWithLatestAndAll } from 'helpers/fiscalYearHelper';
 import { Helmet } from 'react-helmet';
@@ -14,7 +14,11 @@ import Error from 'components/sharedComponents/Error';
 import PageWrapper from 'components/sharedComponents/PageWrapper';
 import { LoadingWrapper } from "components/sharedComponents/Loading";
 import { getBaseUrl, handleShareOptionClick } from 'helpers/socialShare';
-
+import { useHistory } from "react-router-dom";
+import { useQueryParams } from 'helpers/queryParams';
+import { scrollToY } from 'helpers/scrollToHelper';
+import { stickyHeaderHeight } from 'dataMapping/stickyHeader/stickyHeader';
+import { getStickyBreakPointForSidebar } from 'helpers/stickyHeaderHelper';
 import StateContent from './StateContent';
 
 const propTypes = {
@@ -32,11 +36,65 @@ const StatePage = ({
     stateProfile = { fy: '' },
     pickedFy
 }) => {
+    const history = useHistory();
+    const query = useQueryParams();
+    const [activeSection, setActiveSection] = useState(query.section || 'overview');
+
     const slug = `state/${id}/${stateProfile.fy}`;
     const emailArgs = {
         subject: `USAspending.gov State Profile: ${stateProfile.overview.name}`,
         body: `View the spending activity for this state on USAspending.gov: ${getBaseUrl(slug)}`
     };
+
+    const stateSections = [
+        {
+            section: 'overview',
+            label: 'Overview'
+        },
+        {
+            section: 'transactions-over-time',
+            label: 'Transactions Over Time'
+        },
+        {
+            section: 'top-five',
+            label: 'Top 5'
+        }
+    ];
+    const jumpToSection = (section = '') => {
+        // we've been provided a section to jump to
+        // check if it's a valid section
+        const sectionObj = find(stateSections, ['label', section]);
+        if (!sectionObj) return;
+
+        // find the section in dom
+        const sectionDom = document.querySelector(`#state-${sectionObj.section}`);
+        if (!sectionDom) return;
+
+        // add section to url
+        history.replace(`?section=${sectionObj.section}`);
+
+        // add offsets
+        const conditionalOffset = window.scrollY < getStickyBreakPointForSidebar() ? stickyHeaderHeight : 10;
+        const sectionTop = (sectionDom.offsetTop - stickyHeaderHeight - conditionalOffset);
+        scrollToY(sectionTop - 25, 700);
+    };
+
+    useEffect(throttle(() => {
+        // this allows the page to jump to a section on page load, when
+        // using a link to open the page
+        // prevents a console error about react unmounted component leak
+        let isMounted = true;
+        if (isMounted) {
+            const urlSection = query.section;
+            if (urlSection) {
+                setActiveSection(urlSection);
+                jumpToSection(urlSection);
+            }
+        }
+        return () => {
+            isMounted = false;
+        };
+    }, 100), [history, query.section]);
 
     let content = <StateContent id={id} stateProfile={stateProfile} />;
     if (error) {
@@ -71,7 +129,10 @@ const StatePage = ({
                 <ShareIcon
                     onShareOptionClick={handleShare}
                     url={getBaseUrl(slug)} />
-            ]}>
+            ]}
+            sections={stateSections}
+            activeSection={activeSection}
+            jumpToSection={jumpToSection}>
             <main id="main-content" className="main-content">
                 <Helmet>
                     <link href="https://api.mapbox.com/mapbox-gl-js/v2.11.1/mapbox-gl.css" rel="stylesheet" />
